Add files via upload
This code is a script that uses the Keras library to perform data augmentation on a set of images. The script loads all images from a specified directory applies a set of predefined data augmentation techniques using the ImageDataGenerator class and then saves the augmented images to a different directory. The data augmentation techniques include random rotation, shifting, shearing, zooming, and horizontal flipping. The code allows the user to specify the number of augmented images generated for each original image. This script can be helpful in increasing the size of a dataset for training machine learning models.