Stroke and non-Stroke Images Datasets
The Stroke Faces and Non-Strokes Faces Images dataset is a collection of images of individuals who have suffered from a stroke and individuals who have not suffered from a stroke. The images in the dataset are likely to be facial images of the individuals, hence the name "Stroke Faces" and "Non-Strokes Faces". The dataset is likely to be used for training and evaluating machine learning models for the task of classifying an individual as having suffered from a stroke or not based on their facial images